Add legacy tokenizer loading test.
And move tokenizer tests to individual files.
